% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{.svn_control}
\alias{.svn_control}
\title{Check whether a path is under SVN control}
\usage{
.svn_control(path_to_repo = NULL)
}
\arguments{
\item{path_to_repo}{}
}
\value{

}
\description{
Check whether a path is under SVN control
}
