% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn_conflict}
\alias{svn_conflict}
\title{Find conflicts in repo}
\usage{
svn_conflict(path_to_repo = NULL)
}
\arguments{
\item{path_to_repo}{path(s) to the folder(s) whose status to query (if NULL it uses the working directory). Can be a vector.}
}
\value{
character vector of conflicts (tree or otherwise)
}
\description{
Find conflicts in repo
}
\examples{
# svn_conflict(path_to_some_repo)
}
