% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn_log}
\alias{svn_log}
\title{Log of commits in repository/repositories}
\usage{
svn_log(path_to_repo = NULL)
}
\arguments{
\item{path_to_repo}{path(s) to the folder(s) whose status to query (if NULL it uses the working directory). Can be a vector.}
}
\value{
list of dataframes containing repo, revision, user, date, number of lines and commit message
}
\description{
Log of commits in repository/repositories
}
\examples{
# svn_log(path_to_some_repo)
}
