% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn_state}
\alias{svn_state}
\title{States of files in repo}
\usage{
svn_state(path_to_repo = NULL, ignored = FALSE)
}
\arguments{
\item{path_to_repo}{path(s) to the folder(s) whose status to query (if NULL it uses the working directory). Can be a vector.}

\item{ignored}{logical. Return ignored/unversioned files only?}
}
\value{
list containing lists of each type of file status (e.g. added (\code{add}), deleted (\code{delete}), ...)
}
\description{
States of files in repo
}
\examples{
# svn_state(path_to_some_repo)
}
