% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn_details}
\alias{svn_details}
\title{get details on the status of an svn repo}
\usage{
svn_details(path_to_repo = NULL)
}
\arguments{
\item{path_to_repo}{path to the folder whose status to query (if NULL it uses the working directory)}
}
\value{
character
}
\description{
get details on the status of an svn repo
}
\examples{
svn_details()


}
