% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn_revnum}
\alias{svn_revnum}
\title{Get SVN revision number}
\usage{
svn_revnum(path_to_repo = NULL)
}
\arguments{
\item{path_to_repo}{path to repo(s)}
}
\value{
number or list of numbers
}
\description{
Get SVN revision number
}
\examples{
# svn_revnum(path_to_some_repo)
}
